+++ src/locales/ar/messages.json
@@ -34,5 +34,6 @@
   "linkedin": " Linkedin حسابك بموقع",
   "github": "Github حسابك بموقع",
   "twitter": "Twitter حسابك بموقع",
-  "save": "حفظ الملف الشخصي"
+  "save": "حفظ الملف الشخصي",
+  "saved": "تم الحفظ"
 }
